backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
requests==2.31.0
beautifulsoup4==4.12.3
PyMuPDF==1.24.3
pytesseract==0.3.10
Pillow==10.3.0
python-multipart==0.0.9
numpy ==1.26.4
torch==2.3.0
transformers==4.41.0
captum== 0.7.0